test: 删除冗余测试案例
