+++ test_src/test_feat_engineering/test_TargetEncoding.py
@@ -6,18 +6,21 @@ import numpy as np
 from sklearn import datasets
 from src.feat_engineering.TargetEncoding import TargetEncoding
 
-# data generation
-boston = datasets.load_boston()
-x = cate_feat = ['A'] * 200 + ['B'] * 200 + ['C'] * 106
-x = np.array(x)
-y = boston.target
 
-# target encoding
-te = TargetEncoding(num_outer_folds=5, num_inner_folds=5)
-label_pre_train = te.fit(x, y)
+class TestTargetEncoding:
+    def test_fit_transform(self):
+        # data generation
+        boston = datasets.load_boston()
+        x = np.array(['A'] * 200 + ['B'] * 200 + ['C'] * 106)
+        y = boston.target
 
-# obtain target encoded
-label_pre = te.transform(x)
-print(label_pre[:5])
-print(label_pre[201:206])
-print(label_pre[501:506])
+        # target encoding
+        te = TargetEncoding(num_outer_folds=5, num_inner_folds=5)
+        label_pre1 = te.fit_transform(x, y)
+        label_pre2 = te.transform(x)
+        np.testing.assert_array_almost_equal(label_pre1[0], 23.33726682, decimal=2, err_msg="target encoding error")
+        np.testing.assert_array_almost_equal(label_pre1[201], 25.02934193, decimal=2, err_msg="target encoding error")
+        np.testing.assert_array_almost_equal(label_pre1[501], 15.61133666, decimal=2, err_msg="target encoding error")
+        np.testing.assert_array_almost_equal(label_pre2[0], 23.25702477, decimal=2, err_msg="target encoding error")
+        np.testing.assert_array_almost_equal(label_pre2[201], 25.40436784, decimal=2, err_msg="target encoding error")
+        np.testing.assert_array_almost_equal(label_pre2[501], 15.73335422, decimal=2, err_msg="target encoding error")
